Adding new user list format
