+++ lib/glimesh_web/live/user_live/index.html.leex
@@ -1,27 +1,36 @@
 <div class="container">
 
     <h2 class="mt-4"><%= gettext("Our Users") %></h2>
+
+    <form phx-change="search">
+        <input type="text" name="q" value="<%= @query %>" placeholder="Search for usernames" list="results"
+            autocomplete="off" />
+    </form>
+
     <div class="row layout-top-spacing">
         <%= for user <- @users do %>
 
-        <div class="col-md-2 mb-4">
-            <%= live_redirect to: Routes.user_profile_path(@socket, :index, user.username) do %>
+        <div class="col-md-4 mb-4">
             <div class="card">
                 <div class="card-body">
-                    <div class="text-md-center">
-                        <h4><%= user.displayname %></h2>
+                    <h4><%= Glimesh.Chat.Effects.render_username(user) %></h4>
+                    <div class="media">
                         <img src="<%= Glimesh.Avatar.url({user.avatar, user}, :original) %>"
-                            alt="Your Profile Picture"
-                            style="width: 80% !important;"
-                            class="img-fluid mb-2 img-avatar <%= if user.stripe_user_id, do: "img-verified-streamer", else: "" %>">
-
-                        <%= if user.stripe_user_id do %>
-                        <div class="mb-4"><span class="badge badge-secondary"><%= gettext("Sub Ready!") %></span></div>
-                        <% end %>
+                            alt="<%= user.displayname %>" height="128"
+                            class="mr-3 img-avatar <%= if user.stripe_user_id, do: "img-verified-streamer", else: "" %>">
+                        <div class="media-body">
+                            <ul class="list-unstyled">
+                                <li><strong>Joined:</strong> <%= format_datetime(user.inserted_at) %></li>
+                                <li><strong>Following:</strong> 43</li>
+                                <li><strong>Followers:</strong> 142</li>
+                            </ul>
+                            <%= live_redirect to: Routes.user_profile_path(@socket, :index, user.username), class: "stretched-link" do %>
+                            Profile
+                            <% end %>
+                        </div>
                     </div>
                 </div>
             </div>
-            <% end %>
         </div>
 
         <% end %>
